Change guardian email logic a bit
Instead of getting the current email from either User or Guardian
depending on whether Guardian.email is set, from now on we will only
use Guardian.email which should be always set, most of the time
probably being the same as User.email.

I believe this change will make the logic simpler given how the email
needs to function, and it also makes sure the email does not change
even if User.email we get from Tunnistamo changes (which may not be
possible ATM in the practice anyway, but meh).

For this change we need to

* populate existing empty Guardian emails with their User's emails in
  a data migration
* populate empty Guardian email on Guardian save to make sure every
  Guardian's email will be populated