Change unenrolment notification sending
Changed unenrolment notification to be sent only when the unenrolment
is executed using the API unenrolment mutation. Previously it was send
every time an Enrolment object was deleted.

The change was needed, because we don't want to send unenrolment
notifications every time an Enrolment object is deleted, only when
an end user is unenrolling (using the API), or we explicitly delete it
using Django admin UI. An Enrolment object might be deleted also when
the related Occurrence object is deleted, and in that case we want
different notification to be sent.