Add enrolled events past enough to child's past events
In addition to events that have all their occurrences in the past,
pastEvents in ChildNode now contains also events the child has enrolled
in and the enrolled occurrence is in the past enough. There is a
setting for being past enough called
KUKKUU_ENROLLED_OCCURRENCE_IN_PAST_LEEWAY with a default value 30
(mins).