Add ordering for API's "translations" fields
Previously the translations list contained translations in more or
less random order, which made the thing nondeterministic and that
caused problems at least in tests. We could have gotten away with
just by modifying the tests, but that wouldn't have been a completely
trivial task, and anyway IMO it is better to have the list ordered and
return deterministic results.